"int" setPrice changed to "double" setPrice to allign with requirements of price being a double not int

+++ Product.java
@@ -33,7 +33,7 @@ public class Product
         }
     }
 
-    public void setPrice(int setPricein)
+    public void setPrice(double setPricein)
     {
         price = setPricein;
     }
